rename Layout Component to Base
Because:
- to prepare for layout component types
- base makes more sense with the current application of the component
