+++ src/pages/bio.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import Layout from "../components/Layout"
+import Layout from "../components/Base"
 import Bio from "../components/bio"
 import ContentContainer from "../components/ContentContainer"
 const bio = () => {
