+++ src/pages/index.js
@@ -2,7 +2,7 @@ import { Link } from "gatsby"
 import React from "react"
 import Button from "../components/button"
 import Hero from "../components/Hero"
-import TemplateWrapper from "../components/Layout"
+import TemplateWrapper from "../components/Base"
 import SEO from "../components/SEO"
 import styled from "styled-components"
 // import SvgLogo from "../components/Svglogo"
